Be wary of `class`.
I have issues with `class` in JavaScript and avoid it whenever I can:

1. It's misleading. JavaScript has no classical inheritance.

2. It often requires somebody using your class to instantiate it with
the `new` keyword, and it they don't, it can have unintended
consequences.

In this case, I think it's just the wrong abstraction. You aren't using
a constructor, or extending a base class, or doing anything that would
necessitate a class at all. Even more telling, all of your methods are
static. I think what you want is just an object literal containing a few
functions. This, in my experience, is true 95% of the time. Using an
object literal is more _correct_ most of time.